restructure and set up skeleton test
