+++ app/api/api_v1/routers/documents.py
@@ -10,6 +10,8 @@ from fastapi import (
 from app.core.auth import (
     get_current_active_superuser,
 )
+from app.db.crud.document import get_document_and_family
+
 from app.db.crud.deprecated_document import (
     create_document_relationship,
     remove_document_relationship,
@@ -23,6 +25,7 @@ from app.db.crud.deprecated_document import (
 from app.api.api_v1.schemas.document import (
     DocumentDetailResponse,
     DocumentOverviewResponse,
+    DocumentWithFamilyResponse,
     RelationshipAndDocumentsGetResponse,
     RelationshipCreateRequest,
     RelationshipEntityResponse,
@@ -62,7 +65,7 @@ async def document_browse(
 
 @documents_router.get(
     "/documents/{import_id_or_slug}",
-    response_model=DocumentDetailResponse,
+    response_model=Union[DocumentDetailResponse, DocumentWithFamilyResponse],
 )
 async def document_detail(
     import_id_or_slug: str,
@@ -81,6 +84,8 @@ async def document_detail(
     )
     if not group_documents:
         return get_document_detail(db, import_id_or_slug)
+
+    return get_document_and_family(db, import_id_or_slug)
 
 
 @documents_router.post(
